src/emailservice/requirements.txt
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
cryptography==1.7.1
google-api-core==1.2.1
google-auth==1.5.0
google-cloud-core==0.28.1
google-cloud-trace==0.19.0
googleapis-common-protos==1.5.3
grpc-google-iam-v1==0.11.4
grpcio==1.12.1
grpcio-tools==1.12.1
idna==2.7
Jinja2==2.10
keyring==10.1
keyrings.alt==1.3
MarkupSafe==1.0
opencensus==0.1.5
protobuf==3.6.0
pyasn1==0.4.3
pyasn1-modules==0.2.2
pycrypto==2.6.1
pygobject==3.22.0
pytz==2018.5
pyxdg==0.25
requests==2.19.1
rsa==3.4.2
SecretStorage==2.3.1
six==1.11.0
urllib3==1.23
